requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langchain-chroma
streamlit
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
pypdf
arxiv
wikipedia
duckduckgo-search
ddgs
youtube-transcript-api
python-dotenv
watchdog
